requirements.txt
django
django-admin-sortable2
django-ckeditor
dj-database-url
Pillow
psycopg2
pyyaml
